style: move site header into common dir
