+++ src/components/App.jsx
@@ -10,10 +10,9 @@ import { fetchUserAccount, UserAccountApiService } from '@edx/frontend-auth';
 import apiClient from '../config/apiClient';
 import { handleTrackEvents } from '../analytics/analytics';
 import { getLocale, getMessages } from '../i18n/i18n-loader';
-import SiteHeader from './Header/SiteHeader';
+import SiteHeader from './common/SiteHeader';
 import ConnectedProfilePage from './ProfilePage';
 
-import HeaderLogo from '../../assets/edx-sm.png';
 import FooterLogo from '../../assets/edx-footer.png';
 import NotFoundPage from './NotFoundPage';
 
@@ -30,11 +29,7 @@ class App extends Component {
         <Provider store={this.props.store}>
           <ConnectedRouter history={this.props.history}>
             <div>
-              <SiteHeader
-                logo={HeaderLogo}
-                logoDestination={process.env.MARKETING_SITE_BASE_URL}
-                logoAltText={process.env.SITE_NAME}
-              />
+              <SiteHeader />
               <main>
                 <Switch>
                   <Route path="/u/:username" component={ConnectedProfilePage} />
